Add Current location marker to map downloader
This will allow us to see where we are currently, as well as where the projected
location of the rocket is, rather that just where the rocket is.

This means that we will be able to more easily tell where we are meant to be
looking for the rocket, especially when the topographical map that is generated
has very little landmarks, ie roads, building etc.