Implement Baseline GUI structure
Created a baseline GUI layout which uses three seperate tabs to
display the different components of the GUI window.

Implemented the one of the three tabs, the 'Rocket Location' tab
which shows the rockets current location using a Google Static Map
as well as the fetch time, and the option to re-fetch the location
map.

Added placeholder classes for the control and status tabs.

Refactored all documentation to fit the standard python conventions

Refactored project directory strucutre to be more modular