Add manual gps current and target location inputs
The reason for doing this is that; as of yet, we
are unable to actually get the gps location from
the base station. But it still makes sense for us to use
the gui to see the rockets position, and we
could manually get the gps lat and long using the rfm
radio module.